(BOLT-167) Move PuppetDB::Client into bolt proper
