+++ lib/bolt_ext/puppetdb_inventory.rb
@@ -8,66 +8,6 @@ require 'yaml'
 
 module Bolt
   class PuppetDBInventory
-    class Client
-      def self.from_config(config)
-        uri = if config['server_urls'].is_a? String
-                config['server_urls']
-              else
-                config['server_urls'].first
-              end
-        uri = URI.parse(uri)
-        uri.port ||= 8081
-
-        cacert = File.expand_path(config['cacert'])
-        token = config.token
-
-        cert = config['cert']
-        key = config['key']
-
-        new(uri, cacert, token: token, cert: cert, key: key)
-      end
-
-      def initialize(uri, cacert, token: nil, cert: nil, key: nil)
-        @uri = uri
-        @cacert = cacert
-        @token = token
-        @cert = cert
-        @key = key
-      end
-
-      def query_certnames(query)
-        return [] unless query
-
-        body = JSON.generate(query: query)
-
-        response = http_client.post("#{@uri}/pdb/query/v4", body: body, header: headers)
-        if response.code != 200
-          raise "Failed to query PuppetDB: #{response.body}"
-        else
-          results = JSON.parse(response.body)
-          if results.first && !results.first.key?('certname')
-            raise "Query results did not contain a 'certname' field: got #{results.first.keys.join(', ')}"
-          end
-          results.map { |result| result['certname'] }.uniq
-        end
-      end
-
-      def http_client
-        return @http if @http
-        @http = HTTPClient.new
-        @http.ssl_config.set_client_cert_file(@cert, @key)
-        @http.ssl_config.add_trust_ca(@cacert)
-
-        @http
-      end
-
-      def headers
-        headers = { 'Content-Type' => 'application/json' }
-        headers['X-Authentication'] = @token if @token
-        headers
-      end
-    end
-
     class CLI
       def initialize(args)
         @args = args
@@ -142,7 +82,7 @@ query results.
         end
 
         config = Bolt::PuppetDB::Config.new(@config_file, @cli_opts)
-        @puppetdb_client = Client.from_config(config)
+        @puppetdb_client = Bolt::PuppetDB::Client.from_config(config)
 
         unless File.readable?(inventory_file)
           raise "Can't read the inventory file #{inventory_file}"
